requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
pydantic==2.4.2
supabase==1.0.3
motor==3.3.1
python-dotenv==1.0.0
pymongo==4.5.0